Set ChatMode in MessageInputView
