+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/utils/extensions/Channel.kt
@@ -3,25 +3,15 @@ package io.getstream.chat.android.ui.utils.extensions
 import android.content.Context
 import android.text.SpannableString
 import android.text.SpannableStringBuilder
+import com.getstream.sdk.chat.utils.extensions.getUsers
 import io.getstream.chat.android.client.models.Attachment
 import io.getstream.chat.android.client.models.Channel
 import io.getstream.chat.android.client.models.Message
-import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.client.models.name
 import io.getstream.chat.android.livedata.ChatDomain
 import io.getstream.chat.android.ui.R
 import io.getstream.chat.android.ui.channel.list.adapter.ChannelListPayloadDiff
 import io.getstream.chat.android.ui.utils.ModelType
-
-internal fun Channel.getUsers(excludeCurrentUser: Boolean = true): List<User> =
-    members
-        .map { it.user }
-        .let { users ->
-            when {
-                excludeCurrentUser -> users.withoutCurrentUser()
-                else -> users
-            }
-        }
 
 public fun Channel.getDisplayName(context: Context): String =
     name.takeIf { it.isNotEmpty() }
@@ -94,8 +84,6 @@ internal fun Channel.isMessageRead(message: Message): Boolean {
         .mapNotNull { it.lastRead }
         .any { it.time >= message.getCreatedAtOrThrow().time }
 }
-
-internal fun Channel.isDirectMessaging(): Boolean = getUsers().size == 1
 
 // None of the strings used to assemble the preview message are translatable - concatenation here should be fine
 internal fun Channel.getLastMessagePreviewText(
