+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/utils/extensions/User.kt
@@ -7,15 +7,6 @@ import io.getstream.chat.android.client.models.name
 import io.getstream.chat.android.livedata.ChatDomain
 import io.getstream.chat.android.ui.R
 
-internal fun List<User>.withoutCurrentUser(): List<User> {
-    return if (ChatDomain.isInitialized) {
-        val currentUser = ChatDomain.instance().currentUser
-        filter { it.id != currentUser.id }
-    } else {
-        this
-    }
-}
-
 internal fun User.isCurrentUser(): Boolean {
     return if (ChatDomain.isInitialized) {
         id == ChatDomain.instance().currentUser.id
